Refactor to facade/service design pattern
