+++ app/controllers/sessions_controller.rb
@@ -1,8 +1,8 @@
 class SessionsController < ApplicationController
+  def new; end
+
   def create
     require "pry"; binding.pry
-    new_user = UserService.new(params[:something])
-    BackEndFacade.send_request(new_user)
-      # Faraday.post('/api/vi/users/authenticate').with body(new_user)
+    BackEndFacade.authenticate_user('/api/v1/users', params)
   end
 end
