+++ app/facades/back_end_facade.rb
@@ -1,22 +1,7 @@
 class BackEndFacade
   class << self
-    def hello
-      'world'
-    end
-
-    def receive_request(endpoint)
-      conn = Faraday.new(url: endpoint) do |faraday|
-        # faraday.headers['Authorization'] = ENV['bearer']
-      end.get(endpoint)
-      JSON.parse(conn.body)
-    end
-
-    def send_request(endpoint, params)
-      Faraday.post(endpoint) do |faraday|
-        # faraday.headers['Authorization'] = ENV['bearer']
-        faraday.headers['Content-Type'] = 'application/json'
-        faraday.body = params.to_json
-      end
+    def authenticate_user(endpoint, params)
+      BackEndService.send_request(endpoint, params)
     end
   end
 end
