+++ app/views/welcome/index.html.erb
@@ -2,5 +2,5 @@
 <h3>Mission:</h3>
 <p>Find your 'swol mate' and fitness besties by adding your gyms to your profile, adding friends and creating events to go work out.</p>
 <br>
-<%= link_to 'Sign in with Google', login_path, method: :post, class: "btn btn-info btn-sm btn-outline-dark" %>
+<%= link_to 'Sign in with Google', authentication_path, method: :post, class: "btn btn-info btn-sm btn-outline-dark" %>
 <%= link_to 'Create an Account', registration_path, method: :get, class: "btn btn-info btn-sm btn-outline-dark" %>
