--- spec/services/back_end_facade_spec.rb
@@ -1,7 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe BackEndFacade do
-  it 'can call a class method' do
-    expect(BackEndFacade.hello).to eq 'world'
-  end
-end
