+++ spec/services/back_end_service_spec.rb
@@ -1,0 +1,11 @@
+require 'rails_helper'
+
+RSpec.describe BackEndService do
+  it 'can call a class method' do
+    expect(BackEndService.hello).to eq 'world'
+  end
+
+  it 'can do other stuff' do
+    # stuff
+  end
+end
